requirements.txt
python-telegram-bot==20.7
Flask==3.0.0
yt-dlp
ffmpeg-python
gunicorn
